requirements.txt
python-dotenv~=0.15.0
discord
